refactor(identity): move encryption and token services into identity context
- Relocated encryption package into identity bounded context
- Relocated token generation/verification into identity
- Removed shared global dependency
