+++ pkg/identity/tokens/service/create_refresh_token_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/amorindev/go-cms-tmpl/internal/tokens/service"
+	"github.com/amorindev/go-cms-tmpl/pkg/identity/tokens/service"
 	"github.com/stretchr/testify/require"
 )
